añadido factory alumnos y seeder materias

+++ database/migrations/2022_10_09_051958_create_alumnos_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->boolean('rango_tutor');//si es tutor o no
             $table->boolean('cuenta_activa');//estado de cuenta
             $table->integer('semestre');
-            $table->integer('numero_control')->unique();
+            $table->bigInteger('numero_control')->unique();
             $table->string('descripcion')->nullable();
             $table->string('foto_perfil')->nullable();
             $table->timestamps();
